+++ src/Git_Updater/API/Language_Pack_API.php
@@ -10,13 +10,10 @@
 
 namespace Fragen\Git_Updater\API;
 
-use Fragen\Git_Updater\Traits\GU_Trait;
-
 /**
  * Class Language_Pack_API
  */
 class Language_Pack_API extends API {
-	use GU_Trait;
 
 	/**
 	 * Constructor.