+++ src/Git_Updater/Plugin.php
@@ -356,6 +356,10 @@ class Plugin {
 	 * @return mixed
 	 */
 	public function update_site_transient( $transient ) {
+		if ( ! wp_verify_nonce( static::$nonce, 'git-updater' ) ) {
+			return;
+		}
+
 		// needed to fix PHP 7.4 warning.
 		if ( ! \is_object( $transient ) ) {
 			$transient = new \stdClass();
@@ -398,14 +402,12 @@ class Plugin {
 
 				if ( $this->can_update_repo( $plugin ) ) {
 					// Skip on RESTful updating.
-					// phpcs:disable WordPress.Security.NonceVerification.Recommended
 					if ( isset( $_GET['action'], $_GET['plugin'] )
 					&& 'git-updater-update' === $_GET['action']
 					&& $response['slug'] === $_GET['plugin']
 					) {
 						continue;
 					}
-					//phpcs:enable
 
 					// Pull update from dot org if not overriding.
 					if ( ! $this->override_dot_org( 'plugin', $plugin ) ) {
@@ -435,7 +437,6 @@ class Plugin {
 				}
 
 				// Set transient on rollback.
-				// phpcs:ignore WordPress.Security.NonceVerification.Recommended
 				if ( isset( $_GET['plugin'], $_GET['rollback'] ) && $plugin->file === $_GET['plugin']
 				) {
 					$transient->response[ $plugin->file ] = ( new Branch() )->set_rollback_transient( 'plugin', $plugin );